amc done

+++ acm/schema.go
@@ -3,7 +3,7 @@ package acm
 type acmRange struct {
 	StartRange string 
 	EndRange   string 
-	Payment    int    
+	Payment    float64    
 }
 
 var payments = make(map[string][]acmRange)
@@ -13,34 +13,34 @@ func rangePayments(key string) []acmRange {
 		{
 			StartRange: "00:01",
 			EndRange:   "09:00",
-			Payment:    25,
+			Payment:    25.0,
 		},
 		{
 			StartRange: "09:01",
 			EndRange:   "18:00",
-			Payment:    15,
+			Payment:    15.0,
 		},
 		{
 			StartRange: "18:01",
 			EndRange:   "23:00",
-			Payment:    20,
+			Payment:    20.0,
 		},
 	}
 	payments["Weekend"] = []acmRange{
 		{
 			StartRange: "00:01",
 			EndRange:   "09:00",
-			Payment:    30,
+			Payment:    30.0,
 		},
 		{
 			StartRange: "09:01",
 			EndRange:   "18:00",
-			Payment:    20,
+			Payment:    20.0,
 		},
 		{
 			StartRange: "18:01",
 			EndRange:   "23:00",
-			Payment:    25,
+			Payment:    25.0,
 		},
 	}
